+++ src/main/java/org/jabref/logic/importer/CompositeIdFetcher.java
@@ -33,7 +33,7 @@ public class CompositeIdFetcher {
         if (isbn.isPresent()) {
             return new IsbnFetcher(importFormatPreferences)
                     .addRetryFetcher(new EbookDeIsbnFetcher(importFormatPreferences))
-                    .addRetryFetcher(new DoiToBibtexConverterComIsbnFetcher(importFormatPreferences))
+                    // .addRetryFetcher(new DoiToBibtexConverterComIsbnFetcher(importFormatPreferences))
                     .performSearchById(isbn.get().getNormalized());
         }
         /* TODO: IACR is currently disabled, because it needs to be reworked: https://github.com/JabRef/jabref/issues/8876